09/09

+++ src/app/pages/user/detail-product/detail-product.component.ts
@@ -29,9 +29,8 @@ export class DetailProductComponent implements OnInit {
   quantity: number = 1;
   constructor(
     private productService: ProductService,
-    private categoryService: CategoryService,
     private loadingService: LoadingService,
-    private productImageService: ProductImageService,
+    private alertService: AlertService,
     private cartService: CartService,
     private router: ActivatedRoute,
     private route: Router,
@@ -49,56 +48,34 @@ export class DetailProductComponent implements OnInit {
   getProductById(id: number) {
     this.loadingService.show();
     this.productService.getProductById(id).subscribe({
-      next: ({ data, message }) => {
+      next: ({ data }) => {
         this.product = data;
-        this.product.productUrl = `${environment.apiBaseUrl}products/images/${data.thumbnail}`;
-        this.getCategoryById(this.product.category_id);
-        this.getAllProductImagesByProductId(this.productId);
+        this.productImages = [data.thumbnail, ...data.images];
+        this.productImage = this.productImages[0];
       },
-      error: (error) => console.log(error),
+      error: ({ error }) => {
+        this.alertService.error(error.message);
+        this.loadingService.hide();
+      },
       complete: () => this.loadingService.hide(),
     });
   }
-  getCategoryById(id: number) {
-    this.loadingService.show();
-    this.categoryService.getCategoryById(id).subscribe({
-      next: ({ data, message }) => (this.category = data),
-      error: (error) => console.log(error),
-      complete: () => this.loadingService.hide(),
-    });
-  }
-  getAllProductImagesByProductId(productId: number) {
-    this.loadingService.show();
-    this.productImageService
-      .getAllProductImagesByProductId(productId)
-      .subscribe({
-        next: ({ data, message }) => {
-          this.productImages = [
-            this.product.productUrl,
-            ...data.map(
-              (item: ProductImage) =>
-                `${environment.apiBaseUrl}products/images/${item.imageUrl}`
-            ),
-          ];
-          this.productImage = this.productImages[0];
-        },
-        error: (error) => console.log(error),
-        complete: () => this.loadingService.hide(),
-      });
-  }
+
   handleChangeProductImage(productImage: string) {
     this.productImage = productImage;
   }
-  addToCart() {
+  addToCart(): void {
     if (this.tokenService.getToken()) {
       this.cartService.addToCart(this.productId, this.quantity);
     } else {
       this.route.navigate(['login']);
     }
   }
-  navigateToOrderPage() {
+  navigateToOrderPage(): void {
+    if (this.quantity > this.product.quantity) {
+      return;
+    }
     this.loadingService.show();
-
     setTimeout(() => {
       this.orderService.removeAllOrder();
       this.orderService.insertOrderToLocalStorage(
